fix: Moved fetching of best route to separate function
